requirements.txt
Flask==2.3.2
flask-cors==3.0.10
google-genai==0.4.0
dotenv
